Yay proper trending response

+++ src/main/scala/model/Model.scala
@@ -106,5 +106,5 @@ object Engagement {
 
 case class Trending(
     symbol: String,
-    popularity: Int
+    popularity: Float
 )
